feat: create many product features

+++ src/modules/products/services/category/delete-category-service.ts
@@ -1,10 +1,13 @@
 import AppError from '@shared/errors/app-error'
 import ICategoryRepository
   from '@modules/products/infra/repositories/protocols/i-category-repository'
+import ICategoryRelationshipRepository
+  from '@modules/products/infra/repositories/protocols/i-category-relationship-repository'
 
 export default class DeleteCategoryService {
   constructor (
-    private readonly categoryRepository: ICategoryRepository) {}
+    private readonly categoryRepository: ICategoryRepository,
+    private readonly categoryRelationshipRepository: ICategoryRelationshipRepository) {}
 
   public async execute (id?: any): Promise<boolean> {
     if (!id) {
@@ -15,6 +18,13 @@ export default class DeleteCategoryService {
     if (!category) {
       throw new AppError('Category not found')
     }
+
+    const categoryRelationship = await this.categoryRelationshipRepository.findByCategory(category._id)
+    if (!categoryRelationship) {
+      throw new AppError('Category Relatioship not found')
+    }
+
+    await this.categoryRelationshipRepository.delete(categoryRelationship._id)
 
     return await this.categoryRepository.delete(id)
   }
